+++ crates/cli/src/commands/migrate/mod.rs
@@ -4,11 +4,14 @@ mod from_turborepo;
 pub use from_package_json::from_package_json;
 pub use from_turborepo::*;
 
+use crate::helpers::AnyError;
 use moon_workspace::Workspace;
 
-pub async fn check_dirty_repo(workspace: &Workspace) -> Result<(), Box<dyn std::error::Error>> {
+pub async fn check_dirty_repo(workspace: &Workspace) -> Result<(), AnyError> {
     if !workspace.vcs.get_touched_files().await?.all.is_empty() {
-        return Err("Commit or stash your changes before running this command, or use the `--skipTouchedFilesCheck` flag to disable this check.".to_string())?;
+        return Err(
+            "Commit or stash your changes before running this command, or use the `--skipTouchedFilesCheck` flag to disable this check.".into()
+        );
     }
 
     Ok(())